refactor(basin): address review remarks 7-10
- #7 revert the unrelated TokenIssuance reflow so ponder.schema.ts matches
  main (the file is outside the format:check glob; the wrap was noise in an
  already reindex-heavy diff)
- #8 swapDirection moved next to the other leg logic in helpers/basinQuote
  (feeTokenForDirection already lives on BasinFeeService since remark #3)
- #9 shared basinEntityKey/basinEntityInit helpers extract the PK, pool join
  key, and lastUpdatedAt stamps duplicated by BasinDebtService.load and
  BasinFeeService.load; the seed reads stay service-specific
- #10 basin-indexer-rollout.md gets a supersession note pointing at the CFGL
  debt model so nobody builds to the removed swap-redeem batching spec
